backend/pyproject.toml
[project]
name = "daily-stock-backend"
version = "0.1.0"
description = "FastAPI backend for the daily-stock platform"
requires-python = ">=3.10,<4.0"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.6.4",
  "pydantic-settings>=2.2.1",
  "sqlalchemy>=2.0.29",
  "alembic>=1.13.1",
  "psycopg[binary]>=3.1.18",
  "asyncpg>=0.29.0",
  "greenlet>=3.0.3",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.1.1",
  "azure-ai-formrecognizer>=3.3.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.1",
  "mypy>=1.9.0",
  "pytest>=8.1.1",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "RUF"]
ignore = ["E203"]

[tool.ruff.format]
preview = true

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --disable-warnings --cov=app --cov-report=term-missing"

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.mypy]
python_version = "3.10"
strict = true
packages = ["app"]
plugins = []
ignore_missing_imports = false

[tool.poetry]
package-mode = false

[tool.setuptools.packages.find]
include = ["app*"]
